Create README.md
This project is a Social Media Platform built with Django for the backend, Django Templates for the frontend, and MySQL for the database. It features user authentication, where users can sign up and log in using mobile numbers. Users can connect with others, search by name, email, or contact number, and prioritize connected users in search results. The platform supports searching with autocomplete, adds connections, and displays detailed user profiles with connections and mutual connections. It also includes a seamless experience for managing connections and viewing posts.